requirements.txt
aftership==0.2
bottlenose==0.6.3
configure==0.5
ebaysdk==2.1.3
pyyaml==3.11
selenium==2.53.2
requests==2.10.0
